requirements.txt
Flask
openai
requests
python-dotenv
PyMuPDF
scikit-learn
